Move legacy sandbox components to separate folder
