+++ app/scripts/components/sandbox/index.js
@@ -3,20 +3,20 @@ import { Link, Route, Routes, useParams } from 'react-router-dom';
 import styled from 'styled-components';
 import { GridContainer, Grid } from '@trussworks/react-uswds';
 
-import SandboxTypography from './typography';
-import SandboxHug from './hug';
-import SandboxExplorationMap from './exploration-map';
-import SandboxMapBlock from './map-block';
-import SandboxContentBlocks from './content-blocks';
-import SandboxCards from './cards';
-import SandboxMDXPage from './mdx-page';
-import SandboxMDXScrolly from './mdx-scrollytelling';
+import SandboxTypography from './legacy/typography';
+import SandboxHug from './legacy/hug';
+import SandboxExplorationMap from './legacy/exploration-map';
+import SandboxMapBlock from './legacy/map-block';
+import SandboxContentBlocks from './legacy/content-blocks';
+import SandboxCards from './legacy/cards';
+import SandboxMDXPage from './legacy/mdx-page';
+import SandboxMDXScrolly from './legacy/mdx-scrollytelling';
+import SandboxChart from './legacy/mdx-chart';
+import SandboxColors from './legacy/colors';
+import SandboxMDXEditor from './legacy/mdx-editor';
+import SandboxTable from './legacy/table';
+import SandboxLayerInfo from './legacy/layer-info';
 import SandboxOverride from './override';
-import SandboxChart from './mdx-chart';
-import SandboxColors from './colors';
-import SandboxMDXEditor from './mdx-editor';
-import SandboxTable from './table';
-import SandboxLayerInfo from './layer-info';
 import { resourceNotFound } from '$components/uhoh';
 import { Card } from '$components/common/card';
 import PageHero from '$components/common/page-hero';
