+++ app/scripts/components/sandbox/legacy/colors/index.js
