+++ app/scripts/components/sandbox/legacy/content-blocks/index.js
