+++ app/scripts/components/sandbox/legacy/exploration-map/index.js
