+++ app/scripts/components/sandbox/legacy/layer-info/index.js
