+++ app/scripts/components/sandbox/legacy/map-block/index.js
