+++ app/scripts/components/sandbox/legacy/mdx-chart/chart.mdx
