+++ app/scripts/components/sandbox/legacy/mdx-chart/example.csv
