+++ app/scripts/components/sandbox/legacy/mdx-chart/sample.json
