+++ app/scripts/components/sandbox/legacy/mdx-page/index.js
