+++ app/scripts/components/sandbox/legacy/mdx-page/page.mdx
