+++ app/scripts/components/sandbox/legacy/table/index.js
