+++ app/scripts/components/sandbox/legacy/table/table.mdx
